+++ lib/class/xml_data.class.php
@@ -380,7 +380,7 @@ class XML_Data
                     $playlist_name  = $playlist->name;
                     $playitem_total = $playlist->get_media_count('song');
                 } else {
-                    $playlist = new Search(str_replace('smart_', '', (string) $object_id));
+                    $playlist = new Search((int) str_replace('smart_', '', (string) $object_id));
                     $playlist->format();
 
                     $playlist_name  = Search::get_name_byid(str_replace('smart_', '', (string) $object_id));
@@ -639,7 +639,7 @@ class XML_Data
                 $playitem_total = $playlist->get_media_count('song');
                 $playlist_type  = $playlist->type;
             } else {
-                $playlist     = new Search(str_replace('smart_', '', (string) $playlist_id));
+                $playlist     = new Search((int) str_replace('smart_', '', (string) $playlist_id));
                 $playlist->format();
 
                 $playlist_name  = Search::get_name_byid(str_replace('smart_', '', (string) $playlist_id));
